+++ server/package.json
@@ -20,7 +20,10 @@
   },
   "devDependencies": {
     "@types/bcryptjs": "^2.4.6",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
     "@types/node": "^20.10.6",
+    "cors": "^2.8.5",
     "nodemon": "^3.0.2",
     "typescript": "^5.3.3"
   }
